requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
APScheduler>=3.10.0
cryptography>=42.0.0
